moves TodoItem to its own component file - App imports TodoItem
